// WADNR.Web/src/app/shared/generated/model/project-update-contacts-step.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectPersonUpdateItem } from './project-person-update-item';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectUpdateContactsStep { 
    ProjectUpdateBatchID?: number;
    Contacts?: Array<ProjectPersonUpdateItem> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectUpdateContactsStepForm { 
    ProjectUpdateBatchID?: FormControl<number>;
    Contacts?: FormControl<Array<ProjectPersonUpdateItem>>;
}

export class ProjectUpdateContactsStepFormControls { 
    public static ProjectUpdateBatchID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Contacts = (value: FormControlState<Array<ProjectPersonUpdateItem>> | Array<ProjectPersonUpdateItem> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectPersonUpdateItem>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}