// WADNR.Web/src/app/shared/generated/model/project-update-contacts-step-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectPersonUpdateItemRequest } from './project-person-update-item-request';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectUpdateContactsStepRequest { 
    Contacts?: Array<ProjectPersonUpdateItemRequest> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectUpdateContactsStepRequestForm { 
    Contacts?: FormControl<Array<ProjectPersonUpdateItemRequest>>;
}

export class ProjectUpdateContactsStepRequestFormControls { 
    public static Contacts = (value: FormControlState<Array<ProjectPersonUpdateItemRequest>> | Array<ProjectPersonUpdateItemRequest> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectPersonUpdateItemRequest>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}