// WADNR.Web/src/app/shared/generated/model/project-person-update-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectPersonUpdateItem { 
    ProjectPersonUpdateID?: number | null;
    PersonID?: number;
    PersonFullName?: string | null;
    ProjectPersonRelationshipTypeID?: number;
    RelationshipTypeName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectPersonUpdateItemForm { 
    ProjectPersonUpdateID?: FormControl<number>;
    PersonID?: FormControl<number>;
    PersonFullName?: FormControl<string>;
    ProjectPersonRelationshipTypeID?: FormControl<number>;
    RelationshipTypeName?: FormControl<string>;
}

export class ProjectPersonUpdateItemFormControls { 
    public static ProjectPersonUpdateID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static PersonID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static PersonFullName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectPersonRelationshipTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static RelationshipTypeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}